infra/infra/DO/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
jsonpickle==1.0
paramiko==2.4.2
pkg-resources==0.0.0
pyaml==17.12.1
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
python-digitalocean==1.13.2
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
